requirements.txt
Django==5.1.2
djangorestframework==3.15.2
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
django-filter~=24.3
djangorestframework-simplejwt==5.3.1
drf-yasg~=1.21.8
stripe~=11.2.0